backend/requirements.txt
django
psycopg2-binary
django-geojson
dj-database-url
django-leaflet
